pyproject.toml
[tool.poetry]
name = "nano"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {extras = ["voice"], version = "^2.3.2"}
youtube_dl = "^2021.12.17"
geopy = "^2.3.0"
pyyaml = "^6.0.1"
ffmpeg = "^1.4"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.2.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
